refactor(node): extract generic validation to common/, network layer on top
- common/validation.ts: generic createValidationInterceptor({ registry })
- service/validate_response.ts: re-exports common + adds networkRegistry,
  createNetworkValidationInterceptor (network-specific, pre-configured)
- Existing imports from service/validate_response.js keep working
