+++ node/sdk/src/index.ts
@@ -1,9 +1,9 @@
 export * from "./crypto/index.js"
 export * from "./client/client.js"
 export * from "./service/service.js"
+export * from "./common/validation.js"
 export * from "./service/validate_response.js"
 export * from "./service/validate.js"
-export * from "./service/network_validation.js"
 export * from "./service/node.js"
 export { default as NetworkHeaders } from "./common/headers.js"
